requirements.txt
requests
beautifulsoup4
yfinance
google-genai
